Position visualizer's camera and add color background to 3D grids
These changes fix-up the visualizer to look a bit better. The camera now
points to the center of the 3D grid, which better positions the grid
into the center of the view. Thin rectangles have been added behind the
grid on each axis to provide background color, which isn't supported in
the default grid drawing command.