Render visualizer rocket body using a model
By using a model to render the visualizer's rocket body, as opposed to a
simple cone shape using the primitive draw methods, it's possible to
apply a transform to the object.

This sets us up to be able to easily mirror the microcontroller's
gyroscope orientation onto the visualizer's rocket body. This data will
be read in real-time from the MCU's serial bus.