Add driver for BNO055 peripheral with example implementation in app main
These changes add a driver for the BNO055 IMU that is written in c++.

This works excellent, but using a c++ driver has some implications. In
particular, this means that the software entry point must also be
written in c++ and exposed to c externally.

Although the goal was to stick with the C Language, many of the mature
peripheral drivers seem to be written in c++, and so switching the
primary language is a trade-off that seems worthwhile.