Setup flight computer Wifi network connection
These changes add support for Wifi network connection. The current
use-case for this is to synchronize the system clock on boot.

New menu items have been added to `menuconfig` which allow specification
of the network SSID and password. The password is stored in `sdkconfig`,
which is ignored from version control.