Implement link layer for streaming packet transmission and receiving
These changes implement a packet data model which encodes a data payload
for streaming transmission. The payload data is packed into frames where
the boundaries are marked with the byte '~' and reserved characters are
escaped with the byte '}'. The payload is streamed at an interval over
the serial interface.

The visualizer processes the incoming transmission stream by
synchronizing with the boundary markers then parsing the payload
in-between. A partial payload indicates that processing occurred before
the entire payload data was received. The process automatically accounts
for this by simply amending all available data until the end boundary
is found.

This process is capable of self-correcting and appears to be reliable
when tested with small payloads. It's unclear whether the packing scheme
will work for long-range communication.