Stream telemetry data by packing protobuffer into packet payload
These changes finalize a minimal implementation of the flight computer
streaming a telemetry data payload over the serial interface. This is
achieved by using compiled protobuffer schemas to encode and decode the
telemetry data.

The telemetry data is packed into a packet using the existing interface,
allowing for it to be transmitted to the visualizer through wire
protocol.

Next steps will include adding additional payload data and using it in
real-time to update the visualizer.