Read orientation from BNO055 IMU and transmit using protobuf stream
These changes leverage the BNO055 component library to read the device's
orientation. Values are packed into the Telemetry protocol buffer and
transmitted via serial interface to the visualizer.

Useful data is now available to render the scene in real-time.

In the future this will be refactored into a Telemetry task which will
poll data from various sensors and transmit it.