Render visualizer scene using telemetry data stream
These changes refactor the visualizer to both poll the data stream and
render the scene in parallel. This provides a rudimentary working
prototype of the visualizer, where the rocket body is rendered in real
time to match the physical attitude of the flight computer.

There are some noticeable bugs with this implementation. One being that
the visualizer rocket body's center of rotation is at the bottom of the
model rather than at its center. Another issue exists where the
visualizer crashes after some time, likely because it is reading or
writing to memory outside of the buffer's allocated boundaries.