Prevent corrupt heap by allocating additional bytes for packet framing
The packet framing process adds a minimum of two bytes for the packet's
start and stop delimiters, plus an unknown amount of additional bytes
for any escaped data within the packet payload.

Because the framed packet was being written into a buffer the same size
as the packed payload source, we were always writing to unallocated
memory. This caused heap corruption and led to sporadic issues.